LighTDiff/requirements.txt
torch==2.1.2
torchvision==0.16.2
numpy<1.25
scipy<1.12
pillow<11
opencv-python
tqdm
lpips
basicsr>=1.3.4.0
facexlib>=0.2.0.3
lmdb
pyyaml
tb-nightly
yapf
easydict
rawpy
kornia
dominate
trimesh
